extract feature spec search helper from duplicated code in search context and result specs
